Started TS-498 (PeerAgent refactoring).
