+++ src/main/scala-2.11/com/bwsw/tstreams/agents/producer/ProducerMasterServer.scala
@@ -1,0 +1,8 @@
+package com.bwsw.tstreams.agents.producer
+
+/**
+  * Created by ivan on 30.09.16.
+  */
+class ProducerMasterServer {
+
+}
