+++ src/main/scala-2.11/com/bwsw/tstreams/common/ProtocolMessageSerializer.scala
@@ -1,8 +1,8 @@
 package com.bwsw.tstreams.common
 
+import com.bwsw.tstreams.agents.producer.AgentConfiguration
 import com.bwsw.tstreams.coordination.messages.master._
 import com.bwsw.tstreams.coordination.messages.state.{TransactionStateMessage, TransactionStatus}
-import com.bwsw.tstreams.coordination.producer.AgentConfiguration
 
 import scala.collection.mutable
 import scala.util.control.Breaks._
